:building_construction: move SafeInvokeAsync to specialized assembly and new static class; AsyncPatterns
